Set up and successfully implemented first-pass reusable gallery solution
